+++ src/lexers/jsx-lexer.js
@@ -81,6 +81,7 @@ export default class JsxLexer extends JavascriptLexer {
           'warning',
           `"${attributeName}" prop is not a string literal: ${attribute.initializer.expression.text}`
         )
+
         return undefined
       }
 
@@ -91,7 +92,9 @@ export default class JsxLexer extends JavascriptLexer {
 
     const getKey = (node) => getPropValue(node, this.attr)
 
-    if (this.componentFunctions.includes(tagNode.tagName.text)) {
+    if (
+      this.componentFunctions.includes(this.expressionToName(tagNode.tagName))
+    ) {
       const entry = {}
       entry.key = getKey(tagNode)
 
@@ -175,7 +178,12 @@ export default class JsxLexer extends JavascriptLexer {
   }
 
   nodeToString(node, sourceText) {
-    const children = this.parseChildren.call(this, node.children, sourceText)
+    const children = this.parseChildren.call(
+      this,
+      node,
+      node.children,
+      sourceText
+    )
 
     const elemsToString = (children) =>
       children
@@ -209,7 +217,7 @@ export default class JsxLexer extends JavascriptLexer {
       .replace(/(\n|\r)\s*/g, ' ')
   }
 
-  parseChildren(children = [], sourceText) {
+  parseChildren(node, children = [], sourceText) {
     return children
       .map((child) => {
         if (child.kind === ts.SyntaxKind.JsxText) {
@@ -225,13 +233,15 @@ export default class JsxLexer extends JavascriptLexer {
           const name = element.tagName.escapedText
           const isBasic = !element.attributes.properties.length
           const hasDynamicChildren = element.attributes.properties.find(
-            (prop) => prop.name.escapedText === 'i18nIsDynamicList'
+            (prop) =>
+              prop.kind === ts.SyntaxKind.JsxAttribute &&
+              prop.name.escapedText === 'i18nIsDynamicList'
           )
           return {
             type: 'tag',
             children: hasDynamicChildren
               ? []
-              : this.parseChildren(child.children, sourceText),
+              : this.parseChildren(child, child.children, sourceText),
             name,
             isBasic,
             selfClosing: child.kind === ts.SyntaxKind.JsxSelfClosingElement,
@@ -327,7 +337,17 @@ export default class JsxLexer extends JavascriptLexer {
             child.expression.end
           )
 
-          this.emit('warning', `Child is not literal: ${slicedExpression}`)
+          const tagNode = node.openingElement || node
+          const attrValues = tagNode.attributes.properties
+            .filter((attr) => [this.attr, 'defaults'].includes(attr.name?.text))
+            .map(
+              (attr) =>
+                attr.initializer.expression?.text ?? attr.initializer.text
+            )
+
+          if (attrValues.some((attr) => !attr)) {
+            this.emit('warning', `Child is not literal: ${slicedExpression}`)
+          }
 
           return {
             type: 'js',